Move test to integration folder
